Some moar stuff
roomIdle is now a task
